requirements.txt
appnope==0.1.0
asgiref==3.2.10
asn1crypto==1.3.0
backcall==0.2.0
botocore==1.17.43
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
conda==4.8.2
conda-package-handling==1.6.0
cryptography==2.8
decorator==4.4.2
Django==3.1
djangorestframework==3.11.1
docutils==0.15.2
idna==2.8
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pynamodb==4.3.3
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
ruamel-yaml==0.15.87
six==1.14.0
sqlparse==0.3.1
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.2.5
